The type-checked run passes: a wrong dtype and a test that tested the wrong thing
Running the suite with the beartype claw on -- `PIXI_DEV_MODE=1 pixi run -e colsfm
python -m pytest tests/colsfm -q`, which the standard task does not do -- found
five failures. Both causes were real.

* `colsfm.colmap_text_model` declared its COLMAP point colours as
  `Int[ndarray, "n_points 3"]` and returned `uint8`. jaxtyping's `Int` rejects
  unsigned dtypes, so the annotation was simply wrong about the data; `UInt8` is
  what the parser builds and what every reader wants. Four failures.
* `test_a_failed_publication_leaves_the_previous_file_intact` passed a `str`
  where `publish_atomically` declares `bytes` and asserted the resulting
  `TypeError`. Under the claw the violation is raised earlier and is not a
  `TypeError` -- but the deeper problem is that the test never reached the write
  it claims to roll back. It now injects the failure at the *rename*, which is
  the one moment a bad implementation could clobber the destination, and asserts
  both that the good engine survives and that no temporary is left behind.

default suite: 527 passed / 14 skipped. `-m perf`: 7 passed / 534 deselected.
`PIXI_DEV_MODE=1`: 527 passed / 14 skipped. ruff clean.

+++ colsfm/colmap_text_model.py
@@ -26,7 +26,7 @@ from dataclasses import dataclass
 from pathlib import Path
 
 import numpy as np
-from jaxtyping import Float64, Int
+from jaxtyping import Float64, UInt8
 from numpy import ndarray
 from scipy.spatial.transform import Rotation
 
@@ -45,7 +45,7 @@ class ColmapModel:
     """Image name (as written in ``frames_meta.json``) -> camera pose in world."""
     points_xyz: Float64[ndarray, "n_points 3"]
     """Sparse point positions."""
-    points_rgb: Int[ndarray, "n_points 3"]
+    points_rgb: UInt8[ndarray, "n_points 3"]
     """Per-point colour."""
 
 
@@ -105,7 +105,7 @@ def parse_colmap_images_text(text: str) -> list[ColmapImage]:
     return images
 
 
-def parse_colmap_points_text(text: str) -> tuple[Float64[ndarray, "n_points 3"], Int[ndarray, "n_points 3"]]:
+def parse_colmap_points_text(text: str) -> tuple[Float64[ndarray, "n_points 3"], UInt8[ndarray, "n_points 3"]]:
     """Parse the contents of a COLMAP ``points3D.txt`` into positions and colours.
 
     One line per point, so no pairing to get wrong; comments and blank lines are
